requirements.txt
flask
flask-restplus
mysql-connector-python
